[ refactor ] gateway 권한 확인 삭제 및 관련 코드 주석처리(#56 )
