--- gateway/src/main/java/com/linkeleven/msa/gateway/application/dto/UserRoleResponseDto.java
@@ -1,8 +1,0 @@
-package com.linkeleven.msa.gateway.application.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserRoleResponseDto {
-	private String role;
-}
