--- gateway/src/main/java/com/linkeleven/msa/gateway/application/service/UserService.java
@@ -1,7 +1,0 @@
-package com.linkeleven.msa.gateway.application.service;
-
-import com.linkeleven.msa.gateway.application.dto.UserRoleResponseDto;
-
-public interface UserService {
-	UserRoleResponseDto getUserRole(Long userId);
-}
