+++ gateway/src/main/java/com/linkeleven/msa/gateway/client/UserClient.java
@@ -1,0 +1,10 @@
+package com.linkeleven.msa.gateway.client;
+
+/*
+@FeignClient(name="auth-service")
+public interface UserClient {
+
+	@GetMapping("/external/users/{userId}/role")
+	UserRoleResponseDto getUserRole(@PathVariable("userId") Long userId);
+}
+*/
