+++ gateway/src/main/java/com/linkeleven/msa/gateway/client/UserRoleResponseDto.java
@@ -1,0 +1,9 @@
+package com.linkeleven.msa.gateway.client;
+
+/*
+
+@Getter
+public class UserRoleResponseDto {
+	private String role;
+}
+*/
