+++ gateway/src/main/java/com/linkeleven/msa/gateway/filter/CustomPostFilter.java
@@ -1,4 +1,4 @@
-package com.linkeleven.msa.gateway;
+package com.linkeleven.msa.gateway.filter;
 
 import org.springframework.cloud.gateway.filter.GlobalFilter;
 import org.springframework.core.Ordered;
